This is a jupyter notebook for tesseract-OCR and the categorizer.
This module takes as input the address of the folders containing the screenshots. Then the screenshots are passed through Tesseract-OCR system and text from the screenshots are obtained. These texts are then given to the categorizer module which categorizes the text  into one of the following categories:
chat
code
mail
web etc.